Prevent changes to global 'performance' object.
Previously the `Common.now` method polyfilled the `performance.now`
function if it didn't exist but a vendor-prefixed function did exist.

While this might have had a minor performance improvement on subsequent
calls to the function, this was offset by an IIFE that (seemingly?)
served no purpose.

The change also removes the dependency on the global `window` object,
which isn't available when the module is used server-side within a NodeJS
environment.