Remove dependency on `window`.
The Bodies.js file uses references to the `window` object when looking
up the `poly-decomp` library. These references cause errors when using
the module on server-side in a NodeJS environment.

Guarded the references with a `typeof window` check to prevent them
being evaluated in NodeJS. Also included an else branch with an optional
require for the `poly-decomp` module in case it has been installed via
`npm`.